Create Take Home Quiz
This is a take home assignment - assigned on week 5, due week 6. From the  data file, and python file (check the notes folder  in the lesson link - studentRecord.csv, objectExample.py ) complete the class by reading in the gpa,  print out all the student (gpa and name) and faculty instances (name and the research string). Also print out the number of students and the number of faculty.