Create Project 4: Bouncing Shapes
For this project you will learn how to create, move and rotate basic geometric objects in Processing and integrate user input command.  You are to create basic shapes through a key input e.g 'c' creates a circle, 's' creates a square, 't' creates a triangle. Once these shapes are created, give them some movement and a direction. If they run into the edge of the screen, have it bounce off and go another direction. Lastly  if there is a mouseClick within any of the of the shapes, then toggle a color.

 

Point criteria:

 

2 points creating  a  shape

4 creating multiple shapes, having their own variable to manage their movement.

2 points moving a shape in a direction

4 points having a shape bounce off the edges

3 points using mouse click to toggle filling in the shape ( the click must be inside the object)