Ping Pong Final Project-Java
This project continues from the last bouncing shapes project. This is also developed in
Processing. You are to create two paddles and a ball that is able to bounce from the edges and
off the paddles. Each paddle you are able to maneuver up and down within the space with the
selected keyboard keys. This can be a one player (a hand controlling one paddle) or two-player
game with each player to control their paddle sharing the keyboard. The game takes account
the misses and hits by incrementing the misses text box each time the ball passes the paddle
and hits the left/right edge of the screen and incrementing the hit text box each time the ball
hits the paddle.